Move rendering the bundle to a fixture and disable redfish if the collector isn't listed
